Bugfix in Tria, Serializer rewrite and API improvement in Network
Quite a large commit. Tria now handles Q_DECLARE_METATYPE a little bit
better, in the sense that it hopefully doesn't generate code
redeclaring meta types anymore. Serializer in Core was completely
rewritten in a more OOP manner, with more features and based on
MetaObject (Making it easier to use when using Tria). SlotInfo in
Network is now finally a shared structure.