Bugfixes in Core and tests
Mostly a bugfix release for Core, Network and tests.
Core:
 - Removed classes Future and FutureWatcher.

Next up are changes to the build system. The framework will get a Lua
integration using LuaJit. Due to the hassle of getting QMake to build
external dependencies and windows interop issues, we'll use this
opportunity to switch to CMake. Nice thing is that this isn't another
dependency, as it was already needed to build Tria.