Add support to disable Modules via CMake.
Also, Modules can now build against imported versions of their
dependencies. If available the module will be imported from the build tree
(this is the case when you disable a module after building it, without
doing a make clean in between). If there is no version in the build tree,
an installed version will be used. CMake will abort if it finds no
suitable version.

Signed-off-by: Alexander Siol <dunklermeuchler@gmail.com>